refactor: CommentPortal 내의 상태 관련 로직들 훅으로 분리
commentInputModal에서 너무 많은 로직이 존재하여 hook으로 빼기
